GBFS : pas de cache/telemetry pour catch-all 404
