+++ apps/gbfs/test/support/app_config_helper.ex
@@ -3,17 +3,30 @@ defmodule AppConfigHelper do
   A way to temporarily change Application config. Make sure to only use this
   with "async: false" tests
   """
-  import ExUnit.Callbacks, only: [on_exit: 1]
-
-  def change_app_config_temporarily(config_name, config_key, value) do
-    old_value = Application.fetch_env!(config_name, config_key)
-    on_exit(fn -> Application.put_env(config_name, config_key, old_value) end)
-    Application.put_env(config_name, config_key, value)
-  end
-
   def enable_cache do
     change_app_config_temporarily(:gbfs, :disable_page_cache, false)
     # clear the cache since some element can still be there
-    on_exit(fn -> Cachex.clear(:gbfs) end)
+    ExUnit.Callbacks.on_exit(fn -> Cachex.clear(:gbfs) end)
+  end
+
+  def setup_telemetry_handler do
+    events = Transport.Telemetry.gbfs_request_event_names()
+    events |> Enum.at(1) |> :telemetry.list_handlers() |> Enum.map(& &1.id) |> Enum.each(&:telemetry.detach/1)
+    test_pid = self()
+    # inspired by https://github.com/dashbitco/broadway/blob/main/test/broadway_test.exs
+    :telemetry.attach_many(
+      "test-handler-#{System.unique_integer()}",
+      events,
+      fn name, measurements, metadata, _ ->
+        send(test_pid, {:telemetry_event, name, measurements, metadata})
+      end,
+      nil
+    )
+  end
+
+  defp change_app_config_temporarily(config_name, config_key, value) do
+    old_value = Application.fetch_env!(config_name, config_key)
+    ExUnit.Callbacks.on_exit(fn -> Application.put_env(config_name, config_key, old_value) end)
+    Application.put_env(config_name, config_key, value)
   end
 end
